StacNav now uses icons and animated tabbar, cards on home screen added optional prop for custom styling components
