+++ CarboPlanner/App.js
@@ -1,10 +1,8 @@
-import TabNav from "./src/views/TabNav";
 import {StackNav} from "./src/views/StackNav";
 
 
 export default function App() {
   return (
-      //TabNav()
       StackNav()
   );
 }
