+++ CarboPlanner/src/components/Card.tsx
@@ -1,9 +1,14 @@
 import { Children } from 'react';
 import { View, StyleSheet } from 'react-native';
 
-export default function Card({children}) {
+interface CardProps {
+    children: React.ReactNode;
+    customStyle?: object; // Making customStyle prop optional
+  }
+
+export default function Card({children, customStyle}: CardProps) {
     return (
-        <View style={styles.container}>
+        <View style={[styles.container,customStyle]}>
             {children}
         </View>
     );
@@ -16,5 +21,6 @@ const styles = StyleSheet.create({
         paddingHorizontal: '2%',
         marginHorizontal: '2%',
         borderRadius: 25,
+        maxHeight:"100%"
     },
 });
